Move TestAgents to own module
